requirements.txt
django==5.1
psycopg2-binary==2.9.9
pytz==2024.1
celery[redis]==5.4.0
django-celery-beat==2.7.0
python-dotenv==1.0.1
aiogram==3.12.0
asgiref== 3.8.1
aiohttp==3.10.5
pillow==10.4.0
prometheus_client==0.20.0
gunicorn==23.0.0
django-storages==1.14.4
boto3==1.35.22
